+++ package.json
@@ -1,4 +1,8 @@
 {
+  "scripts": {
+    "dev": "next dev"
+  },
+  
   "dependencies": {
     "next": "^14.1.0",
     "react": "^18.2.0",
